Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -8,7 +8,7 @@ int main() {
     sum=a+b;
     diff=a-b;
     mul=a*b;
-    div=int(a/b);
+    div=float(a/b);
     if(ch=='+'){
         printf("%d",sum);
     }
